AppGlobConst: Add guard to password validators to handle NULL values
When the Change Password form, or any other form for that matter, is 
reset via Angular's Form.reset() function, the form's control values 
will be set to `null`. Generally, we would expect the value of a textbox 
control to be an empty string, so when these values are null, it can 
cause unexpected behavior. As a result, all validators added by 
BAASU-380 now invoke a function that verifies if a value is NULL and 
returns an empty string (if so) or the value of the field (if non-NULL).