ReqDetail: Convert Title Bar of dialog window to non-flex layout
Flex is a good approach for making responsive designs that can work 
across many devices, but this is a case of it being misused. In this 
case, because we could not float a flex element, we were attempting to 
offset the flex basis of various elements so that they would line up 
correctly. While that worked when we explicitly set the Material Dialog 
pane's format/width, that design breaks when the dialog is resized. The 
fact that this design relied on a fixed width (specified in pixels) for 
an element is an anti-pattern for Flex designs that violates the 
fundamental principles of Flex.