BAASU-329: (PR) Include Client-Side ID in Request & expect in Response
This allows us to associate MatFileUploadComponent objects with 
appropriate SavedRequestModel objects. (Prior to this commit, we 
operated on the assumption that 
MatFileUploadComponent.queueData().toArray() will always return objects 
in a consistent order. While we have never observed that assumption 
being violated, I shared a concern expressed during PR, so I offered to 
modify this code on the front-end as well as making the necessary 
backend changes to ensure that this does not become a problem down the 
road.)

This set of changes only handles the inclusion of those IDs in the 
requests; however, this merge request does NOT perform the associations.